+++ login-workflow/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.3.0
+### Added
+-   i18n utilities
+    -   This was relocated from [@pxblue/react-auth-shared](https://www.npmjs.com/package/@pxblue/react-auth-shared)
+-   DismissKeyboardView component
+    -   This was relocated from @pxblue/react-auth-shared
+
+
 ## v1.2.1
 ### Changed
 -   Shareable logic has been extracted to [@pxblue/react-auth-shared](https://www.npmjs.com/package/@pxblue/react-auth-shared). This package re-exports all exports from the shared package, so there are no breaking changes in usage. The majority of the API documentation in now in the shared repository.